[Annotation] move service registration to own Extension
