+++ src/DI/ApiGenExtension.php
@@ -2,8 +2,6 @@
 
 namespace ApiGen\DI;
 
-use ApiGen\Annotation\AnnotationDecorator;
-use ApiGen\Contracts\Annotation\AnnotationSubscriberInterface;
 use ApiGen\Contracts\Generator\GeneratorInterface;
 use ApiGen\Contracts\Generator\GeneratorQueueInterface;
 use Latte\Engine;
@@ -26,7 +24,6 @@ final class ApiGenExtension extends CompilerExtension
     public function beforeCompile(): void
     {
         $this->setupGeneratorQueue();
-        $this->setupAnnotationDecorator();
     }
 
     private function setupTemplating(): void
@@ -45,17 +42,6 @@ final class ApiGenExtension extends CompilerExtension
             GeneratorQueueInterface::class,
             GeneratorInterface::class,
             'addGenerator'
-        );
-    }
-
-    // @todo: decouple to own package
-    private function setupAnnotationDecorator(): void
-    {
-        DefinitionCollector::loadCollectorWithType(
-            $this->getContainerBuilder(),
-            AnnotationDecorator::class,
-            AnnotationSubscriberInterface::class,
-            'addAnnotationSubscriber'
         );
     }
 }
