+++ src/config/config.neon
@@ -1,5 +1,6 @@
 extensions:
     - ApiGen\DI\ApiGenExtension
+    - ApiGen\Annotation\DI\AnnotationExtension
     - ApiGen\Element\DI\ElementExtension
     - ApiGen\Reflection\DI\ReflectionExtension
     - ApiGen\ModularConfiguration\DI\ModularConfigurationExtension
